+++ agrest-jaxrs2-openapi/src/main/java/io/agrest/jaxrs2/openapi/parmconverter/UriInfoResolver.java
@@ -4,6 +4,7 @@ import io.agrest.jaxrs2.openapi.TypeWrapper;
 import io.agrest.protocol.ControlParams;
 import io.agrest.protocol.Direction;
 import io.swagger.v3.jaxrs2.ResolvedParameter;
+import io.swagger.v3.oas.models.Operation;
 import io.swagger.v3.oas.models.media.Schema;
 import io.swagger.v3.oas.models.media.StringSchema;
 import io.swagger.v3.oas.models.parameters.Parameter;
@@ -12,9 +13,41 @@ import io.swagger.v3.oas.models.parameters.QueryParameter;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.UriInfo;
 import java.lang.annotation.Annotation;
+import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 public class UriInfoResolver {
+
+    /**
+     * Remove UriInfo parameters from the Operation if they were also explicitly added.
+     */
+    static void removeDuplicatedParams(Operation operation) {
+
+        List<Parameter> params = operation.getParameters();
+        if(params == null) {
+            return;
+        }
+
+        // remove duplicate parameters that may have been added via UriInfo
+        Set<String> nonUriInfoParams = new HashSet<>();
+        List<Parameter> uriInfoParams = new ArrayList<>();
+
+        for (Parameter p : params) {
+            if (p instanceof UriInfoQueryParameter) {
+                uriInfoParams.add(p);
+            } else {
+                nonUriInfoParams.add(p.getName());
+            }
+        }
+
+        for (Parameter p : uriInfoParams) {
+            if (nonUriInfoParams.contains(p.getName())) {
+                params.remove(p);
+            }
+        }
+    }
 
     public boolean willResolve(TypeWrapper wrapper, List<Annotation> annotations) {
         if (wrapper != null && wrapper.getRawClass() == UriInfo.class) {
@@ -95,8 +128,16 @@ public class UriInfoResolver {
     }
 
     protected Parameter queryParam(ControlParams param) {
-        return new QueryParameter()
+        // returning our own subclass of the QueryParameter to be able to analyze duplicate parameters
+        // downstream
+        return new UriInfoQueryParameter()
                 .name(param.name())
                 .description(param.description);
+    }
+
+    /**
+     * A subclass indicating that this is an implicit parameters added via UriInfo
+     */
+    static class UriInfoQueryParameter extends QueryParameter {
     }
 }